refactor exceptions.
